feat: Allow images to be embedded from ChurchTools
This massively reduces disk usage on the WordPress installation by not saving event images to local storage.

It utilizes a filter provided by the Events Manager plugin to override the image URL with the URL to the image in ChurchTools.
This URL is stored in a custom event attribute whose name has to be set in the plugin's settings to enable this feature.